don't link to index.htmls anymore
there's no need to! it was only for when browsing the files locally, but
it's much recommended to use an http server for this anyway thanks to
icon svgs apparently only working offline. also this means we are only
supporting hosted servers, which gives us room to mess around with local
storage for per-browser settings, if that's a thing we want to do.