don't duplic8 contrib&artist trax on artist pages!
if an artist is both an artist and contributor for a track, that track
would show up twice on the artist page, and with the exact same line,
due to the way the artist track list is constructed. this started
happening because we combine the "as contributor" and "as artist" arrays
for a given artist, which can have overlap! so now we throw a unique()
utility function on top and it works like its supposed to.