html: clone tags & templates passed in via slots
I'm not 100% sure the right behavior here in the long run (whether
we should be doing a deep clone or not), so for now I haven't added
any specific tests. Snapshot tests covering uses of templates which
depend on cloning (i.e. parents which reuse a given template) will
do better to make sure everything keeps working like it should.