data, yaml: inherit music-related properties from original release
When a track has 'Originally Released As', these fields are now
automatically inherited:

* Artists
* Contributors
* Referenced Tracks
* Sampled Tracks

Including any of these fields alongside 'Originally Released As'
is an error.

Corresponding properties are valid, but ignored.

This uses a new "compositional" style to define how each of these
properties inherits while retaining the original behavior for
tracks that aren't re-releases, and avoids hard-coding much of
anything!