data: always define composite utilities with `key() {}` syntax
Sublime Text doesn't index the key in `key: () => {}` as a symbol
for function definitions if the parameter list takes up more than
one line, but always works for `key() {}`.

This also just makes it a little easier to add "preamble" before
the main return value, when relevant. Consistent syntax is usually
a plus for recurring behavioral forms!