linkExternal: add content slot
There's not necessarily a use for linkExternal with content slot
on its own as of this commit, compared to a normal html.tag('a'),
because linkExternal basically just wraps formatExternalLink and
the content slot opts out of using that. But this does allow this
component to be used in places now... and extended upon... later!