client: fix minor dynamic tooltip placement math error
This code is meant to replace a rectangle that is the result of
an intersection with a spanning-infinitely-left region, i.e.
the opportunity area for placing a tooltip to the left of its
hoverable, with a narrow channel so that its `x` is the
appropriate left alignment for the tooltip.

That behavior is working correctly (hopefully), but the fallback
is that it *shouldn't* be modifying a rectangle which intersected
with a spanning-infinitely-right region. We were mistakenly
dropping anything to do with the intersection and returning
the region rectangle instead, which broke vertical alignment
any time a tooltip is aligned to the right of the hoverable.