data: withReverse{Reference,Contribution}List: better exiting
An existing test for withAlbum (which is semi-newly based on
withUniqueReferencingThing, and thus withReverseReferenceList)
indicates that if we just can't *find* the album in a provided
data array, that's not an exit condition. This commit adapts
withReverseReferencingList to respect that when the data list
is empty.

(Treating an empty array as the same thing as an array that
 isn't provided at all - the existing behavior - is really just
 a bug!)