content, css: use flex layout to bottom-align commentary dates
If the shadow effect is useful as part of the cue, it needs to
be there every time.

This prevents the main text from entering the same horizontal
column as the date. That's generally OK but can result in a small
main column in "sometime 6/12/2012 - 8/5/2012" shaped dates.
We can try grid or non-binary flex-grow/shrink later, if wanted.