css: set line-height on sticky heading h1
Apparently manually setting line-height resolves some
instability between a line of text which includes JP
characters and one which does not - it appears that
without a manual line-height, the line height is
computed independently per line. This mostly just
causes an apparent shift when collapsing down to
one line. Avoid it completely.