css: conclude collapse anim w/ height: auto
Permit the browser to take over height. This should always be
the same value as we animated to, but avoids unpredicted cases
just cutting off part of the heading completely. This has an
impact on changing the font size (but not pixel size) while
the sticky heading is collapsed, for example.