data: Track: withHasUniqueCoverArt: process direct links only
No soupyFind/soupyReverse.

This commit introduces a notion of difference between
an artwork which is 'declared to exist' via, say, having
cover artists, versus one which... is 'usable', because it
has resolved cover artists. No other code addressing this
yet, so it's in a tense state where existence is in every
case according to declared cover artists, whether or not
those cover artist references resolve.