data: withConstitutedArtwork: don't pass find through
The whole point of constituting artworks in advance is so that
they exist with a stable identity having been collected into
artworkData during save() from various data loading steps.
(The new data check reportOrphanedArtworks confirms this at
runtime.) We already included artworkData in linkWIkiDataSpec
for exactly this purpose, so just drop the 'find' dependency
and transfer here.

As a result, constitutibleArtworkList has no dependencies of
its own, and dependencies on coverArtworks, trackArtworks, etc
come only from earlier steps, i.e. deciding whether or not to
constitute an artwork in the first place.